requirements.txt
# Core Django
Django==5.0.7
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication & JWT
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7

# CORS Headers
django-cors-headers==4.3.1

# Filtering
django-filter==24.2

# Redis & Caching
redis==5.0.4
django-redis==5.4.0
hiredis==2.3.2

# Production Server
gunicorn==22.0.0

# Environment & Configuration
python-dotenv==1.0.1

# Development Tools
django-debug-toolbar==4.3.0

# Testing
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Code Quality
black==24.4.2
flake8==7.0.0
isort==5.13.2
pylint==3.2.2
pylint-django==2.5.5

# Type Checking
mypy==1.10.0
django-stubs==5.0.2

# Security
django-ratelimit==4.1.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
